+++ Shadow Adventure/src/entities/player.cpp
@@ -116,14 +116,14 @@ void PLAYER::setCHAR(char ch) // set character of player
 
 bool PLAYER::voidUnder(std::vector<std::vector<block>>& board) // if player hit the floor
 {
-	if (board[x[x.size() - 1] + 1][y[y.size() - 1]].character == '#' || 
-		board[x[x.size() - 1] + 1][y[y.size() - 1]].character == '|' || 
-		board[x[x.size() - 1] + 1][y[y.size() - 1]].character == '-' ||  
-		board[x[x.size() - 1] + 1][y[y.size() - 1]].character == 'S' ||
-		board[x[x.size() - 1] + 1][y[y.size() - 1]].character == '+' ||
-		board[x[x.size() - 1] + 1][y[y.size() - 1]].character == 'M' ||
-		board[x[x.size() - 1] + 1][y[y.size() - 1]].character == character ||
-		board[x[x.size() - 1] + 1][y[y.size() - 1]].character == 'F') return true;
+	if (board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == '#' ||
+		board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == '|' ||
+		board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == '-' ||
+		board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == 'S' ||
+		board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == '+' ||
+		board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == 'M' ||
+		board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == character ||
+		board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == 'F') return true;
 
 	return false;
 }
@@ -131,25 +131,25 @@ bool PLAYER::collisions(std::vector<std::vector<block>>& board, int type) // if 
 {
 	if (type == 80) 
 	{
-		if (board[x[x.size() - 1] + 1][y[y.size() - 1]].character == '#' || 
-			board[x[x.size() - 1] + 1][y[y.size() - 1]].character == 'F' || 
-			board[x[x.size() - 1] + 1][y[y.size() - 1]].character == '+' ||
-			board[x[x.size() - 1] + 1][y[y.size() - 1]].character == 'S' ||
-			board[x[x.size() - 1] + 1][y[y.size() - 1]].character == character) return true;
+		if (board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == '#' ||
+			board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == 'F' ||
+			board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == '+' ||
+			board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == 'S' ||
+			board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == character) return true;
 	}
 	for (int dot = 0; dot < x.size() || dot < y.size(); dot++) // check every player's element
 	{
-		if (type == 77) if (board[x[dot]][y[dot] + 1].character == '#' || 
-							board[x[dot]][y[dot] + 1].character == 'D' || 
-							board[x[dot]][y[dot] + 1].character == 'F' || 
-							board[x[dot]][y[dot] + 1].character == 'S' ||
-							board[x[dot]][y[dot] + 1].character == character || y[0] == board[0].size() - 2) return true;
+		if (type == 77) if (board[x[dot]][y[dot] + (uint64_t)1].character == '#' ||
+							board[x[dot]][y[dot] + (uint64_t)1].character == 'D' ||
+							board[x[dot]][y[dot] + (uint64_t)1].character == 'F' ||
+							board[x[dot]][y[dot] + (uint64_t)1].character == 'S' ||
+							board[x[dot]][y[dot] + (uint64_t)1].character == character || y[0] == board[0].size() - 2) return true;
 
-		if (type == 75) if (board[x[dot]][y[dot] - 1].character == '#' || 
-							board[x[dot]][y[dot] - 1].character == 'D' || 
-							board[x[dot]][y[dot] - 1].character == 'F' || 
-							board[x[dot]][y[dot] - 1].character == 'S' ||
-							board[x[dot]][y[dot] - 1].character == character || y[0] == 1) return true;
+		if (type == 75) if (board[x[dot]][y[dot] - (uint64_t)1].character == '#' ||
+							board[x[dot]][y[dot] - (uint64_t)1].character == 'D' ||
+							board[x[dot]][y[dot] - (uint64_t)1].character == 'F' ||
+							board[x[dot]][y[dot] - (uint64_t)1].character == 'S' ||
+							board[x[dot]][y[dot] - (uint64_t)1].character == character || y[0] == 1) return true;
 	}
 	return false;
 }
@@ -165,12 +165,12 @@ bool PLAYER::ladder(std::vector<std::vector<block>>& board) // check player on l
 }
 bool PLAYER::limit(std::vector<std::vector<block>>& board) // if the player is in prohibited territory
 {
-	if (board[x[x.size() - 1] + 1][y[y.size() - 1]].character == 'D') return true;
+	if (board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == 'D') return true;
 	return false;
 }
 bool PLAYER::onspeedbooster(std::vector<std::vector<block>>& board) // if player stay on speed booster element
 {
-	if (board[x[x.size() - 1] + 1][y[y.size() - 1]].character == 'S') return true;
+	if (board[x[x.size() - 1] + (uint64_t)1][y[y.size() - 1]].character == 'S') return true;
 	return false;
 }
 bool PLAYER::onstairs(std::vector<std::vector<block>>& board, int type) // if player stay on stairs
@@ -179,14 +179,14 @@ bool PLAYER::onstairs(std::vector<std::vector<block>>& board, int type) // if pl
 	{
 		for (int dot = 0; dot < x.size() || dot < y.size(); dot++) // check every player's element
 		{
-			if (board[x[dot]][y[dot] + 1].character == '+') return true;
+			if (board[x[dot]][y[dot] + (uint64_t)1].character == '+') return true;
 		}
 	}
 	if (type == 75)
 	{
 		for (int dot = 0; dot < x.size() || dot < y.size(); dot++) // check every player's element
 		{
-			if (board[x[dot]][y[dot] - 1].character == '+') return true;
+			if (board[x[dot]][y[dot] - (uint64_t)1].character == '+') return true;
 		}
 	}
 	return false;
